+++ app/src/main/java/com/w36495/senty/domain/error/ErrorMessageProvider.kt
@@ -1,0 +1,5 @@
+package com.w36495.senty.domain.error
+
+interface ErrorMessageProvider {
+    fun getMessageRes(throwable: Throwable?): Int
+}
