+++ app/src/main/java/com/w36495/senty/view/screen/main/MainViewModel.kt
@@ -1,25 +1,15 @@
 package com.w36495.senty.view.screen.main
 
 import androidx.lifecycle.ViewModel
-import com.w36495.senty.R
-import com.w36495.senty.data.error.AuthError
-import com.w36495.senty.data.error.ErrorHandler
-import com.w36495.senty.data.error.GlobalError
+import com.w36495.senty.domain.error.ErrorMessageProvider
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 
 @HiltViewModel
 class MainViewModel @Inject constructor(
-    private val errorHandler: ErrorHandler,
+    private val errorMessageProvider: ErrorMessageProvider,
 ) : ViewModel() {
-    fun getErrorMessage(throwable: Throwable?): Int {
-        val errorType = errorHandler.handleError(throwable)
-        return when (errorType) {
-            AuthError.ExistedEmail -> R.string.error_signup_exsist_email
-            AuthError.FailedLogin -> R.string.error_login_failed
-            AuthError.TooManyRequest -> R.string.error_login_failed_too_many_request
-            GlobalError.NetWork -> R.string.common_error_message_network
-            GlobalError.UnKnownError -> R.string.common_error_unknown
-        }
+    fun getErrorMessageRes(throwable: Throwable?): Int {
+        return errorMessageProvider.getMessageRes(throwable)
     }
 }
